Quiz creation update
I'm storing the number of questions in the quiz in a session variable called $quesno and then running a loop to enter the quiz questions and options one by one. When a user enters a quiz question, it will redirect to quiz4.php and redirect back to quiz3.php until all questions are recorded in the database.